refactor(tabbar): Move tabbar component in components directory & rename user profil screen
