+++ src/components/tabBar/TabBar.tsx
@@ -1,12 +1,12 @@
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
 import CreateMemeScreen from '@screens/meme/CreateMemeScreen';
+import UserProfileScreen from '@screens/profile/UserProfile';
+import ProfileSvg from 'assets/profile';
 import React from 'react';
 
-import CreateSvg from './../../../assets/create';
-import HomeSvg from './../../../assets/home';
-import ProfileSvg from './../../../assets/profile';
-import FeedScreen from './../feed/feedScreen';
-import ProfileScreen from './../profile/profileScreen';
+import CreateSvg from '../../../assets/create';
+import HomeSvg from '../../../assets/home';
+import FeedScreen from '../../screens/feed/feedScreen';
 
 const Tab = createBottomTabNavigator();
 
@@ -21,7 +21,7 @@ const TabBar = () => {
 						iconComponent = <HomeSvg size={size} fill={focused ? 'blue' : 'black'} />;
 					} else if (route.name === 'CreateMeme') {
 						iconComponent = <CreateSvg size={size} fill={focused ? 'blue' : 'black'} />;
-					} else if (route.name === 'Profile') {
+					} else if (route.name === 'UserProfile') {
 						iconComponent = (
 							<ProfileSvg size={size} fill={focused ? 'blue' : 'black'} />
 						);
@@ -37,7 +37,11 @@ const TabBar = () => {
 				component={CreateMemeScreen}
 				options={{ headerShown: false }}
 			/>
-			<Tab.Screen name="Profile" component={ProfileScreen} options={{ headerShown: false }} />
+			<Tab.Screen
+				name="UserProfile"
+				component={UserProfileScreen}
+				options={{ headerShown: false }}
+			/>
 		</Tab.Navigator>
 	);
 };
