+++ src/index.tsx
@@ -10,10 +10,10 @@ import { Text } from 'react-native';
 import { Provider } from 'react-redux';
 import { PersistGate } from 'redux-persist/integration/react';
 
+import TabBar from './components/tabBar/TabBar';
 import LoginScreen from './screens/auth/LoginScreen';
 import SignupScreen from './screens/auth/SignupScreen';
 import VerificationScreen from './screens/auth/VerificationScreen';
-import TabBar from './screens/tabBar/tabBar';
 
 const { store, persistor } = setupStore();
